Now the library will remember what the user did or did not click on.
The way this is done is very expensive and should be cleaned up, but this may require chaning how the arrays are structured.

In addition, it broke the feature of selecting the last option.

The ability to remember what is selected should eventually be a option that can be turned on/off to help with performance